Solve 1688

+++ rust/examples/1688_v1.rs
@@ -2,7 +2,7 @@ use std::error::Error;
 use std::fmt::Write;
 use std::io::{stdin, Read};
 
-fn get_ccw(p1: &Vector, p2: &Vector, p3: &Vector) -> usize {
+fn get_ccw(p1: &Vector, p2: &Vector, p3: &Vector) -> i64 {
     let Vector(x1, y1) = p1;
     let Vector(x2, y2) = p2;
     let Vector(x3, y3) = p3;
@@ -20,10 +20,10 @@ fn get_polygon_inside(polygon_v: &Vec<Vector>, point: &Vector) -> bool {
             return true;
         }
         if (point.1 > polygon_v[idx].1) != (point.1 > polygon_v[jdx].1) {
-            let crossed_x: f64 = ((polygon_v[jdx].0 - polygon_v[idx].0)
-                * (point.1 - polygon_v[idx].1)
-                / (polygon_v[jdx].1 - polygon_v[idx].1)
-                + polygon_v[idx].0) as f64;
+            let crossed_x: f64 = (polygon_v[jdx].0 - polygon_v[idx].0) as f64
+                * (point.1 - polygon_v[idx].1) as f64
+                / (polygon_v[jdx].1 - polygon_v[idx].1) as f64
+                + polygon_v[idx].0 as f64;
             if crossed_x > point.0 as f64 {
                 count_crossed += 1;
             }
@@ -31,12 +31,12 @@ fn get_polygon_inside(polygon_v: &Vec<Vector>, point: &Vector) -> bool {
     }
     count_crossed % 2 != 0
 }
-struct Vector(usize, usize);
+struct Vector(i64, i64);
 fn main() -> Result<(), Box<dyn Error>> {
     let mut output = String::new();
     let mut input = String::new();
     stdin().read_to_string(&mut input).unwrap();
-    let mut input = input.split_ascii_whitespace().flat_map(str::parse::<usize>);
+    let mut input = input.split_ascii_whitespace().flat_map(str::parse::<i64>);
     let num_v = input.next().unwrap() as usize;
     let mut polygon_v = Vec::new();
     for _ in 0..num_v {
